refactor(Footer): add JSDoc, improve structure, rename file, removes inline comments
