+++ src/components/Footer/Footer.jsx
@@ -1,11 +1,14 @@
 import { Link } from 'react-router-dom';
 import { CiInstagram, CiFacebook, CiLinkedin } from 'react-icons/ci';
 
+/**
+ * Footer component containing site links, social media links, and acknowledgements.
+ * @returns {JSX.Element} - Rendered Footer component.
+ */
 export default function Footer() {
   return (
     <footer className="bg-bg-muted py-10 text-gray-800">
       <div className="mx-auto flex max-w-7xl flex-col items-center justify-between px-6 md:flex-row md:items-start">
-        {/* Logo */}
         <div className="mb-6 text-center md:mb-0 md:text-left">
           <div className="logo mb-2 font-logo text-2xl">
             <Link to="/" className="hover:text-bg-highlight">
@@ -14,8 +17,6 @@ export default function Footer() {
           </div>
           <p className="text-xs">© {new Date().getFullYear()} Holidaze.</p>
         </div>
-
-        {/* Navigation Links */}
         <div className="mb-6 md:mb-0">
           <h3 className="mb-2 text-lg">Quick Links</h3>
           <ul className="space-y-1 text-center">
@@ -41,8 +42,6 @@ export default function Footer() {
             </li>
           </ul>
         </div>
-
-        {/* Social Media Icons */}
         <div className="mb-6 text-center md:mb-0">
           <h3 className="mb-2 text-lg">Follow Us</h3>
           <div className="flex justify-center space-x-4 text-2xl">
@@ -73,11 +72,7 @@ export default function Footer() {
           </div>
         </div>
       </div>
-
-      {/* Divider Line */}
       <div className="mx-auto my-6 max-w-7xl border-t border-gray-300 px-6"></div>
-
-      {/* Copyright & Acknowledgement */}
       <div className="mb-2 px-6 text-center">
         <p className="text-sm">
           <strong>Acknowledgement:</strong> This website is not real and does not represent a real
