+++ src/components/Layout/index.jsx
@@ -1,5 +1,5 @@
 import { Outlet, useLocation } from 'react-router-dom';
-import Footer from '../Footer';
+import Footer from '../Footer/Footer';
 import Header from '../Header';
 
 export default function Layout() {
